basically finished need dummy data, succulent add if want(last thing)

+++ plantEdit.php
@@ -2,7 +2,11 @@
 
 include 'classes/foliage.class.php';
 
-$_SESSION['plantid'] = $plantid = $_SERVER['QUERY_STRING'];
+if(isset($_SESSION['plantid'])){
+  unset($_SESSION['plantid']);
+}
+
+$plantid = $_SERVER['QUERY_STRING'];
 ?>
 
 <!DOCTYPE html>
@@ -18,8 +22,8 @@ $_SESSION['plantid'] = $plantid = $_SERVER['QUERY_STRING'];
           integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <!-- custom css -->
     <link rel="stylesheet" type="text/css" href="assets/style/style.css">
-    <title>"My G"reens</title>
-  </head>
+    <title>"My G" || Edit</title>
+    </head>
 
   <body>
     <header>
@@ -27,39 +31,48 @@ $_SESSION['plantid'] = $plantid = $_SERVER['QUERY_STRING'];
     </header>
     <!-- END HEADER -->
     <!-- PLANT PROFILE -->
-    <?php 
-    $plantdata = file_get_contents("anitaAllFoliage.json");
-    $plantdata = json_decode($plantdata, true);
-    $_SESSION['pnn'] = $plantdata[$plantid]['pnn'];
-    $_SESSION['img'] = $img = $plantdata[$plantid]['img'];
-  ?>
-    <div class="container">
-      <div class="name-tag">
-        <h3 id="intro-name"><?php echo $_SESSION['pnn']; ?></h3>
+    <?php
+      $plantdata = file_get_contents("anitaAllFoliage.json");
+      $plantdata = json_decode($plantdata, true);
+      $_SESSION['pnn'] = $plantdata[$plantid]['pnn'];
+      $_SESSION['img'] = $img = $plantdata[$plantid]['img'];
+      $_SESSION['plantid'] = $plantid;
+    ?>
+    <div class="container-fluid">
+
+      <div class="name-tag mb-5 mb-md-0">
+        <h3 class="tag">Edit</h3>
+        <div id="mpp" class="name-tag-name"><?php echo $_SESSION['pnn']; ?></div>
       </div>
-      <div class="container">
+
+      <div class="container-fluid mt-md-5">
         <div class="row">
-          <div class="col">
-            <img src="<?php echo $_SESSION['img']?>" class="plant-prof">
+          <div class="col text-center">
+
+            <img src="<?php echo $_SESSION['img']?>" class="plant-prof-pic plant-prof float-none float-md-right" >
           </div>
-          <div class="col-md-6 pb-2 pb-md-0">
-            <form method="post" action="dhPlantEditF.php" enctype="multipart/form-data">
-              <div class="formLine text-center text-md-left my-lg-10">
+          <div class="col-md-6 d-flex flex-column align-items-center align-items-md-start">
+            <form class="mt-3" method="post" action="dhPlantEditF.php" enctype="multipart/form-data">
+
+              <div class="form-group">
                 <label for="photo">Edit <?php echo $_SESSION['pnn']; ?>'s picture</label><br>
-                <input type="file" name="photo">
-              </div>
-              <div class="formLine text-center text-md-left my-lg-10">
+                <input class="form-control" type="file" name="photo">
+</div>
+<div class="form-group">
                 <label>Change <?php echo $_SESSION['pnn']; ?>'s name</label><br>
-                <input type="text" name="pnn" value="<?php echo $_SESSION['pnn']; ?>" placeholder="<?php echo $_SESSION['pnn']; ?>">
-              </div>
+                <input class="form-control" type="text" name="pnn" value="<?php echo $_SESSION['pnn']; ?>" placeholder="<?php echo $_SESSION['pnn']; ?>">
+</div>
               <input type="hidden" name="plantid" value="<?php echo $_SESSION['plantid']; ?>">
-              <input type="submit" name="update" class="update-btn" value="Update">
+
+              <input type="submit" name="update" class="cbtn update-btn mt-md-4" value="Save Changes">
             </form>
 
-            <button type="button" class="delete-btn" data-toggle="modal"
+            <div class="d-modal mt-auto">
+            <button type="button" class="cbtn delete-btn" data-toggle="modal"
                     data-target="#deleteplant<?php echo $_SESSION['plantid']; ?>">
               Delete <?php echo $_SESSION['pnn']; ?>
             </button>
+
             <div class="modal fade" id="deleteplant<?php echo $_SESSION['plantid']; ?>" tabindex="-1" role="dialog"
                  aria-labelledby="deleteplantLabel<?php echo $_SESSION['plantid']; ?>" aria-hidden="true">
               <div class="modal-dialog" role="document">
@@ -69,13 +82,14 @@ $_SESSION['plantid'] = $plantid = $_SERVER['QUERY_STRING'];
                     <button type="button" class="close" data-dismiss="modal" aria-label="Close">
                       <span aria-hidden="true">&times;</span>
                     </button>
+
                   </div>
                   <div class="modal-body">
                     Are you sure you want to delete this plant?
                   </div>
                   <div class="modal-footer">
-                    <a href="allFoliage.php" class="n-btn">No.</a>
-                    <a href="dhFoliageDelete.php?<?php echo $_SESSION['plantid']; ?>" class="delete-btn">Yes,
+                    <a href="allFoliage.php" class="cbtn update-confirm-btn">No.</a>
+                    <a href="dhFoliageDelete.php?<?php echo $_SESSION['plantid']; ?>" class="cbtn delete-confirm-btn">Yes,
                       delete!</a>
                   </div>
                 </div>
@@ -84,8 +98,9 @@ $_SESSION['plantid'] = $plantid = $_SERVER['QUERY_STRING'];
               </div>
               </div>
               </div>
+</div>
 
-       
+
 
 
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
